add search product name

+++ src/main/java/com/example/phoneshop/dao/BrandDAO.java
@@ -7,7 +7,7 @@ import com.example.phoneshop.model.Brand;
 
 public interface BrandDAO {
 	List<Brand> get();
-	Brand get(int id);
+	Brand get(long id);
 	void save(Brand brand);
 	void delete(int id);
 }
